// Front-end/OfferteApp/src/interfaces/Materiaalsoort.ts
export interface Material {
    id: number
    name: string;
    spatrand: number;
    vensterbank: number;
    boorgatenPerStuk: boolean;
    wcd: boolean;
    randafwerking: boolean;
    prijsPerM2: number;
    randafwerkingPerM: number;
    spatrandPerM: number;
    vensterbankPerM: number;
    uitsparingOnderbouw: number;
    uitsparingInleg: number;
    uitsparingRuw: number;
    kraangat: number;
    zeepdispenser: number;
    boorgatenPerStukPrijs: number;
    wcdPrijs: number;
    achterwandPerM: number;
    randafwerkingPerMPrice: number;
}